refactor: revert to wiki-compliant page naming with prefix system
