+++ src/pages/_components/ExercisesSection.tsx
