+++ src/pages/_components/HeroSection.tsx
