+++ src/pages/exercises_Ósemka.tsx
