+++ src/pages/footer_Ustawienia.tsx
