+++ src/pages/other_Autorzy.tsx
